+++ backend/controllers/EventoController.php
@@ -2,11 +2,15 @@
 
 namespace backend\controllers;
 
+use common\models\CartaSearch;
 use common\models\Evento;
 use common\models\EventoSearch;
+use common\models\User;
+use yii\filters\AccessControl;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\web\Response;
 
 /**
  * EventoController implements the CRUD actions for Evento model.
@@ -21,6 +25,19 @@ class EventoController extends Controller
         return array_merge(
             parent::behaviors(),
             [
+                'access' => [
+                    'class' => AccessControl::class,
+                    'rules' => [
+                        [
+                            'actions' => ['index', 'view', 'escolher_carta', 'escolher_carta_update', 'create', 'update', 'delete'],
+                            'allow' => true,
+                            'roles' => ['admin'],
+                        ],
+                    ],
+                    'denyCallback' => function ($rule, $action) {
+                        return $this->redirect(['site/index']);
+                    }
+                ],
                 'verbs' => [
                     'class' => VerbFilter::className(),
                     'actions' => [
@@ -34,10 +51,14 @@ class EventoController extends Controller
     /**
      * Lists all Evento models.
      *
-     * @return string
+     * @return string|Response
      */
     public function actionIndex()
     {
+        if (!\Yii::$app->user->can('readEvento')) {
+            return $this->redirect(['site/index']);
+        }
+
         $searchModel = new EventoSearch();
         $dataProvider = $searchModel->search($this->request->queryParams);
 
@@ -51,23 +72,51 @@ class EventoController extends Controller
      * Displays a single Evento model.
      * @param int $id ID
      * @param int $carta_id Carta ID
-     * @return string
+     * @return string|Response
      * @throws NotFoundHttpException if the model cannot be found
      */
     public function actionView($id, $carta_id)
     {
+        if (!\Yii::$app->user->can('readEvento')) {
+            return $this->redirect(['site/index']);
+        }
+
         return $this->render('view', [
             'model' => $this->findModel($id, $carta_id),
         ]);
     }
 
     /**
+     * O admin escolhe uma carta para associar a um novo evento
+     * @return string|Response
+     */
+    public function actionEscolher_carta()
+    {
+        if (!\Yii::$app->user->can('createEvento')) {
+            return $this->redirect(['site/index']);
+        }
+
+        $searchModel = new CartaSearch();
+        $dataProvider = $searchModel->search($this->request->queryParams);
+
+        return $this->render('escolher_carta', [
+            'searchModel' => $searchModel,
+            'dataProvider' => $dataProvider,
+        ]);
+    }
+
+    /**
      * Creates a new Evento model.
      * If creation is successful, the browser will be redirected to the 'view' page.
+     * @param int $id Carta ID
      * @return string|\yii\web\Response
      */
-    public function actionCreate()
+    public function actionCreate($id)
     {
+        if (!\Yii::$app->user->can('createEvento')) {
+            return $this->redirect(['site/index']);
+        }
+
         $model = new Evento();
 
         if ($this->request->isPost) {
@@ -80,6 +129,28 @@ class EventoController extends Controller
 
         return $this->render('create', [
             'model' => $model,
+            'carta_id' => $id,
+        ]);
+    }
+
+    /**
+     * O admin escolhe uma carta para associar a um novo evento
+     * @return string|Response
+     */
+    public function actionEscolher_carta_update($id, $carta_id)
+    {
+        if (!\Yii::$app->user->can('updateEvento')) {
+            return $this->redirect(['site/index']);
+        }
+
+        $searchModel = new CartaSearch();
+        $dataProvider = $searchModel->search($this->request->queryParams);
+
+        return $this->render('escolher_carta_update', [
+            'searchModel' => $searchModel,
+            'dataProvider' => $dataProvider,
+            'id' => $id,
+            'carta_id' => $carta_id,
         ]);
     }
 
@@ -88,11 +159,16 @@ class EventoController extends Controller
      * If update is successful, the browser will be redirected to the 'view' page.
      * @param int $id ID
      * @param int $carta_id Carta ID
+     * @param int $carta_nova Carta ID nova
      * @return string|\yii\web\Response
      * @throws NotFoundHttpException if the model cannot be found
      */
-    public function actionUpdate($id, $carta_id)
+    public function actionUpdate($id, $carta_id, $carta_nova)
     {
+        if (!\Yii::$app->user->can('updateEvento')) {
+            return $this->redirect(['site/index']);
+        }
+
         $model = $this->findModel($id, $carta_id);
 
         if ($this->request->isPost && $model->load($this->request->post()) && $model->save()) {
@@ -101,6 +177,7 @@ class EventoController extends Controller
 
         return $this->render('update', [
             'model' => $model,
+            'carta_nova' => $carta_nova,
         ]);
     }
 
@@ -114,6 +191,10 @@ class EventoController extends Controller
      */
     public function actionDelete($id, $carta_id)
     {
+        if (!\Yii::$app->user->can('deleteEvento')) {
+            return $this->redirect(['site/index']);
+        }
+
         $this->findModel($id, $carta_id)->delete();
 
         return $this->redirect(['index']);
